update golang queries

+++ backend/golang/internal/infra/services_enriched_queries.sql.go
@@ -19,9 +19,9 @@ JOIN "specialists" "sp" ON "s"."specialist_id" = "sp"."id"
 JOIN "service_names" "sn" ON "s"."service_name_id" = "sn"."id"
 JOIN "specializations" "sz" ON "sn"."specialization_id" = "sz"."id"
 WHERE true
-   AND ($1::text = '' OR LOWER(unaccent("sp"."name")) LIKE '%' || LOWER(unaccent($1)) || '%')
+   AND ($1::text = ''     OR LOWER(unaccent("sp"."name")) LIKE '%' || LOWER(unaccent($1)) || '%')
    AND ($2::text = '' OR LOWER(unaccent("sz"."name")) LIKE '%' || LOWER(unaccent($2)) || '%')
-   AND ($3::text = '' OR LOWER(unaccent("sn"."name")) LIKE '%' || LOWER(unaccent($3)) || '%')
+   AND ($3::text = ''    OR LOWER(unaccent("sn"."name")) LIKE '%' || LOWER(unaccent($3)) || '%')
 `
 
 type CountServicesEnrichedParams struct {
@@ -47,11 +47,11 @@ JOIN "specialists" "sp" ON "s"."specialist_id" = "sp"."id"
 JOIN "service_names" "sn" ON "s"."service_name_id" = "sn"."id"
 JOIN "specializations" "sz" ON "sn"."specialization_id" = "sz"."id"
 WHERE true
-   AND ($1::text = '' OR LOWER(unaccent("sp"."name")) LIKE '%' || LOWER(unaccent($1)) || '%')
+   AND ($1::text = ''     OR LOWER(unaccent("sp"."name")) LIKE '%' || LOWER(unaccent($1)) || '%')
    AND ($2::text = '' OR LOWER(unaccent("sz"."name")) LIKE '%' || LOWER(unaccent($2)) || '%')
-   AND ($3::text = '' OR LOWER(unaccent("sn"."name")) LIKE '%' || LOWER(unaccent($3)) || '%')
-LIMIT $5
-OFFSET $4
+   AND ($3::text = ''    OR LOWER(unaccent("sn"."name")) LIKE '%' || LOWER(unaccent($3)) || '%')
+LIMIT $5::integer
+OFFSET $4::integer
 `
 
 type ListServicesEnrichedParams struct {
